ai-services/requirements.txt
anthropic>=0.7.0
openai>=1.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
python-dotenv>=1.0.0
google-cloud-vision>=3.4.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.20.0
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
flask>=2.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
aiofiles>=0.23.0
